requirement.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
aws-requests-auth==0.4.3
billiard==3.6.4.0
boto3==1.17.100
botocore==1.20.100
celery==5.0.5
certifi==2021.5.30
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2
django-after-response==0.2.2
django-celery-beat==2.2.0
django-cors-headers==3.5.0
django-environ==0.4.5
django-filter==2.4.0
django-heroku==0.3.1
django-timezone-field==4.1.2
django-upload-validator==1.1.5
djangorestframework==3.12.2
djangorestframework-jwt==1.0.1
djangorestframework-simplejwt==4.6.0
drf-yasg==1.20.0
drfdocs==0.0.11
filelock==3.4.2
googlemaps==4.5.3
gunicorn==20.1.0
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.1.0
libmagic==1.0
MarkupSafe==2.0.1
msgpack==1.0.2
packaging==20.9
prompt-toolkit==3.0.19
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.0.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.5
python-crontab==2.5.1
python-dateutil==2.8.1
python-dotenv==0.19.2
python-magic==0.4.22
python-magic-bin==0.4.14
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.2
s3transfer==0.4.2
service-identity==21.1.0
six==1.16.0
sms==0.4
sqlparse==0.4.1
swagger-spec-validator==2.7.3
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing-extensions==3.10.0.2
uritemplate==3.0.1
urllib3==1.26.5
validate-email==1.3
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
zope.interface==5.4.0
django-xhtml2pdf==0.0.4
reportlab==3.6.6
pytest
Faker
whitenoise==5.2.0
dj-database-url==0.5.0